requirements.txt
# Core Testing Framework
playwright==1.41.0
pytest==8.0.0
pytest-playwright==0.4.4
pytest-xdist==3.5.0
pytest-rerunfailures==13.0

# Reporting
allure-pytest==2.13.2
pytest-html==4.1.1

# Utilities
python-dotenv==1.0.1
faker==22.6.0
pyyaml==6.0.1

# Data Management
openpyxl==3.1.2
pandas==2.2.0

# Logging
loguru==0.7.2

# API Testing
requests==2.31.0
jsonschema==4.21.1
requests-toolbelt==1.0.0

# Mobile Testing - Appium
Appium-Python-Client==3.1.1
selenium==4.17.2

# Soft Assertions
pytest-assume==2.4.3

# Email Reporting
secure-smtplib==0.1.1
premailer==3.10.0
jinja2==3.1.3
